pyproject.toml
[tool.poetry]
name = "parivaha"
version = "0.1.5"
description = ""
authors = ["Suhail <suhailece@gmail.com>"]
readme = "README.md"
packages = [{include = "parivaha", from = "src"}]

[tool.poetry.dependencies]
python            = "^3.11.7"
click             = "^8.1"
python-frontmatter= "^1.1"
notionmanager     = "^0.1.24"
python-dotenv     = "^1.0"
tqdm              = "^4.67.1"

[tool.poetry.scripts]
parivaha = "parivaha.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
